Adding logic for receiving registration result.
Bug fixes:
* Reverting to 32bit uint index as compute shaders do not allow 16bit uint.
* Logic errors in sending of data.

+++ HoloIntervention/Content/SpatialMapping/SurfaceMesh.h
@@ -52,7 +52,7 @@ namespace HoloIntervention
 
     struct IndexBufferType
     {
-      uint16 index;
+      uint32 index;
     };
 
     struct OutputBufferType
